[Part 2] Rename modules to use snake_case (#99)
Rename modules to use snake_case and remove unused files in xscape. Remove xscape code related to interactive prompt.

* Rename commonAnalytic.py to common_analytic.py

* Rename reconcileEvents.py to reconcile_events.py

* Rename plotcostsAnalytic.py to  plotcosts_analytic.py

* Remove unused newickFormatReader.py, getInput.py
